Add back duplicate filenames under separate sections.
This is particularly important for the TNT config, where the "canonical"
filename for some of the tracks is not obvious. The logic in Chocolate
Doom allows duplicate entries for the same hash, so we'll check each
entry to see if we can find the referenced file. But keeping the dupes
in a separate section allows music pack authors to determine the minimum
number of tracks which need to be generated to cover a game.