Dedup duplicate tracks.
Some of Doom's music tracks are reused for multiple levels. In some
cases these different lumps have the same hash but there are several
where the hash is different even though the music is the same. So,
reduce the set of filenames in the Doom 1 config file, based on the
data from this page:

  <https://doomwiki.org/wiki/Doom_music>

This should allow smaller music packs to be produced - the config file
lists exactly which files should be included to cover all of Doom's
music, without any ambiguity.